Renaming files
Renaming files to better reflect their function
